requirements.txt
# Whereas install_requires requirements are minimal, requirements files often contain an
# exhaustive listing of pinned versions for the purpose of achieving repeatable
# installations of a complete environment.

setuptools~=65.5.0
werobot~=1.13.1
fastapi~=0.88.0
uvicorn
passlib~=1.7.4
pydantic~=1.10.2

databases~=0.7.0
sqlalchemy~=1.4.45
pyjwt~=2.4.0
pymysql~=1.0.2